Refactor/balance update

+++ src/main/java/com/tech/padawan/financialmanager/transaction/dto/CreateTransactionDTO.java
@@ -6,12 +6,13 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public record CreateTransactionDTO(
         @NotNull(message = "Value is required")
         @Min(value = 0, message = "Value most be a positive number")
-        double value,
+        BigDecimal value,
         String description,
         @NotNull(message = "Category is required")
         Long categoryId,
