+++ src/main/java/com/tech/padawan/financialmanager/user/dto/UserSearchedDTO.java
@@ -3,6 +3,7 @@ package com.tech.padawan.financialmanager.user.dto;
 import com.tech.padawan.financialmanager.role.model.RoleType;
 import com.tech.padawan.financialmanager.user.model.User;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public record UserSearchedDTO(
@@ -10,7 +11,7 @@ public record UserSearchedDTO(
         String name,
         String email,
         Date birthdate,
-        double balance,
+        BigDecimal balance,
         RoleType role
 ) {
     public static UserSearchedDTO from(User user) {
